SIM-2874: Clean up stricter mypy issues and move SAGAAgent to actions.ActionsAgent basically.
